Student Class

+++ src/main/java/org/example/Calculator.java
@@ -4,7 +4,7 @@ public class Calculator {
     public double a;
     public double b;
 
-    public Calculator(int a, int b) {
+    public Calculator(double a, double b) {
         this.a = a;
         this.b = b;
     }
